requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
pubg-python==0.12.5


#
acachecontrol==0.3.5
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.2.1
async-generator==1.10       
async-timeout==4.0.2        
attrs==21.4.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.23.10
botocore==1.26.10
Cerberus==1.3.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-config-url==0.1.1
Django==4.0.4
django-cors-headers==3.11.0
django-csp==3.7
django-oauth-toolkit==1.7.0
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-social-oauth2==1.2.1
drf-yasg==1.20.0
ecdsa==0.17.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
frozenlist==1.3.0
furl==2.1.3
future==0.18.2
h11==0.13.0
idna==3.3
inflection==0.5.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.0
json5==0.9.10
jsonschema==4.5.1
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
jwcrypto==1.4.2
lml==0.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
mysqlclient==2.1.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.2
oauthlib==3.2.2
okta-jwt-verifier==0.2.3
openapi==1.1.0
openpyxl==3.0.10
orderedmultidict==1.0.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.3
pubg-python==0.12.5
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2021.3
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
selenium==4.4.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
tcell-agent==1.8.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.9
validator==0.7.1
virtualenv==20.16.3
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
websocket-client==1.4.2
workbook==1.1
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yarl==1.7.2